I switched the code to run on jacobian, vector products
This should improve runtime. It also dramatically simplifies the code.

I also moved the SSE code + forward method to another file... I'll come back to these later, but don't want to worry about them right now. Currently, getting this thing running with an arbitrary right hand side + arbitrary number of equations is the priority.
